Move remaining REST tests across to new test structure
