+++ spec/common/globals/named_dependencies.js
@@ -11,5 +11,6 @@ define(function() {
     // test modules
     'globals':           { browser: 'spec/common/globals/environment',    node: 'spec/common/globals/environment' },
     'shared_helper':     { browser: 'spec/common/modules/shared_helper',  node: 'spec/common/modules/shared_helper' },
+    'async':             { browser: 'node_modules/async/lib/async' }
   };
 });
