+++ spec/support/browser_setup.js
@@ -52,6 +52,9 @@ require([(baseUrl + '/spec/common/globals/named_dependencies.js').replace('//','
       },
       'browser-base64': {
         exports: 'Base64'
+      },
+      'async': {
+        exports: 'async'
       }
     },
